Add support for two factor authentication.
The users will now be able to activate this on their accounts,
in a new security panel in account preferences. This generates
a random secret and displayes a QR code that the user can store
on their device.

They are also able to generate backup codes (each can be used once)

The admin can also disallow this feature by setting allow_2fa
to false in the Trac configuration file.